Thunks: Move PackedArguments to a dedicated header
